Moved Authentication service registration in AuthenticationServices and renamed project namespaces and csproj file for consistency https://github.com/achobanov/web-application/issues/3
